Consensus class has been renamed and moved in vcg lib. Code has been changed accordingly to import and use the new class. I took the chance to delete some unused stuff as .h, typedefs, functions etc...

+++ src/meshlabplugins/filter_feature_alignment/filter_feature_alignment.h
@@ -81,7 +81,7 @@ class FilterFeatureAlignment : public QObject, public MeshFilterInterface
         static typename ALIGNER_TYPE::Result RigidTransformationOperation(MeshModel& mFix, MeshModel& mMov, typename ALIGNER_TYPE::Parameters& param, CallBackPos *cb=NULL);
 
         template<class CONSENSUS_TYPE>
-        static int ConsensusOperation(MeshModel& mFix, MeshModel& mMov, typename CONSENSUS_TYPE::Parameters& param, CallBackPos *cb=NULL);
+        static float ConsensusOperation(MeshModel& mFix, MeshModel& mMov, typename CONSENSUS_TYPE::Parameters& param, CallBackPos *cb=NULL);
 
         template<class ALIGNER_TYPE>
         static typename ALIGNER_TYPE::Result RansacOperation(MeshModel& mFix, MeshModel& mMov, typename ALIGNER_TYPE::Parameters& param, CallBackPos *cb=NULL);
